✨ feat: ID/PW 로그인·회원가입 추가 (#643) (#647)
* refactor: memberId Int → String 마이그레이션 (Phase A of #643)

- ChallengerInfo.memberId, MemberOAuth.memberId 등 Auth/Home/MyPage 도메인 모델을 String 으로 통일
- MemberOAuthDTO/MyProfileDTO/ChallengerMemeberDTO 등 서버 응답 매핑을 String 기준으로 정합화
- AppStorageKey 에 memberIdString/legacyMemberIdInt 어댑터 헬퍼 추가
- Activity/Notice 등 Int 를 유지하는 호출부에는 경계 변환(Int(...) ?? 0) 적용
- SignUpViewModelTests / Mock / Preview 데이터 String 타입으로 정렬

* feat: ID/PW 로그인·회원가입 API 레이어 추가 (Phase B of #643)

- AuthRouter 에 loginByIdPw / registerByIdPw / checkLoginIdAvailability 엔드포인트 추가
- 신규 DTO: LoginByIdPwRequest/Response, RegisterByIdPwRequest/Response, CheckLoginIdAvailabilityResponse
- 신규 도메인 모델: LoginByIdPwResult, RegisterByIdPwResult (memberId String + TokenPair)
- AuthRepositoryProtocol/AuthRepository 에 3개 메서드 추가 + Mock 구현 (reviewer/umc12345! 시나리오)
- LoginByIdPwUseCase / RegisterByIdPwUseCase / CheckLoginIdAvailabilityUseCase 신설 + Provider 등록

* feat: ID/PW 로그인·회원가입 화면 추가 (Phase C of #643)

- AppFlow에 showSignUpByIdPw 콜백, AppState에 signUpByIdPw 케이스 추가
- LoginView 상단에 ID/PW 입력 섹션 + "회원가입" 진입 텍스트 버튼 추가
  - LoginByIdPwSection 컴포넌트 신설 (TextField/SecureField + 인라인 에러 + MainButton + 구분선)
- LoginViewModel: loginIdInput/passwordInput/loginByIdPwState/loginByIdPwErrorMessage 추가
  - loginWithIdPw() 액션: 401(서버 메시지) 인라인 노출, 네트워크 오류는 ErrorHandler Alert 위임
- SignUpByIdPwView 신설 (단일 ScrollView 6 섹션: 이메일/ID/비밀번호×2/이름·닉네임/학교/약관)
- SignUpByIdPwViewModel 신설:
  - 로그인 ID 500ms debounce 중복 검사 + 결과 캐시(loginIdAvailabilityCache)
  - 비밀번호 8자/일치, 이름 ≤10자, 닉네임 한글 1~5자, 필수 약관 검증
  - 가입 성공 시 fetchMyProfile로 승인 여부 판정해 main/pendingApproval 분기

* test: ID/PW Auth UseCase 단위 테스트 추가 (Phase D of #643)

- LoginByIdPwUseCaseTests: 로그인 성공 시 토큰 저장, 실패 시 에러 전파 + 토큰 미저장
- RegisterByIdPwUseCaseTests: 가입 성공 시 토큰 즉시 저장, 실패 시 에러 전파 + 토큰 미저장
- CheckLoginIdAvailabilityUseCaseTests: true/false 통과, 에러 전파
- StubAuthRepository (actor) + InMemoryTokenStore (actor) 공용 mock 정의

+++ AppProduct/AppProduct/Features/MyPage/Data/DTO/MyPageProfileDTO.swift
@@ -252,7 +252,7 @@ extension MyPageProfileResponseDTO {
             .flatMap { UMCPartType(apiValue: $0) } ?? .admin
 
         let challengerInfo = ChallengerInfo(
-            memberId: id.intValue,
+            memberId: id,
             gen: latestRecord?.gisu.intValue ?? latestRole?.gisu?.intValue ?? 0,
             name: latestRecord?.name ?? name,
             nickname: latestRecord?.nickname ?? nickname,
